Track quantity; Remove only one per item

+++ pantry/app/page.js
@@ -1,7 +1,7 @@
 'use client'
 import {Box, Stack, Typography, Button, Modal, TextField} from '@mui/material'
 import {firestore} from '@/firebase' 
-import {collection, query, getDocs, doc, setDoc, deleteDoc} from 'firebase/firestore'
+import {collection, query, getDocs, doc, setDoc, deleteDoc, getDoc} from 'firebase/firestore'
 import {useEffect, useState} from 'react'
 
 const style = {
@@ -33,7 +33,7 @@ export default function Home() {
     const docs = await getDocs(snapshot)
     const pantryList = []
     docs.forEach((doc) => {
-      pantryList.push(doc.id)
+      pantryList.push({"name": doc.id, ...doc.data()})
     }) 
     console.log(pantryList)
     setPantry(pantryList)
@@ -45,13 +45,31 @@ export default function Home() {
 
   const addItem = async (item) => {
     const docRef = doc(collection(firestore, 'pantry'), item)
-    await setDoc(docRef, {})
+    // Check if exists
+    const docSnap = await getDoc(docRef)
+    if (docSnap.exists()) {
+      const {count} = docSnap.data()
+      await setDoc(docRef, {count: count + 1})
+    } 
+    else {
+      console.log(docSnap.data())
+      await setDoc(docRef, {count: 1})
+    }
     await updatePantry()
   }
 
   const removeItem = async (item) => {
     const docRef = doc(collection(firestore, 'pantry'), item)
-    await deleteDoc(docRef)
+    const docSnap = await getDoc(docRef)
+    if (docSnap.exists()) {
+      const {count} = docSnap.data()
+      if (count === 1) {
+        await deleteDoc(docRef)
+      }
+      else {
+        await setDoc(docRef, {count: count - 1})
+      }
+    }
     await updatePantry()
   }
 
@@ -109,9 +127,9 @@ export default function Home() {
 
         <Stack width="800px" height="300px" spacing={2} overflow="auto">
 
-          {pantry.map((i) => (
+          {pantry.map(({name, count}) => (
               <Box
-                key={i}
+                key={name}
                 width="100%"
                 minHeight="150px"
                 display={"flex"}
@@ -122,10 +140,15 @@ export default function Home() {
               >
                 <Typography variant={'h3'} color={'#333'} textAlign={'center'}>
                     {
-                      i.charAt(0).toUpperCase() + i.slice(1) 
+                      name.charAt(0).toUpperCase() + name.slice(1) 
                     }
                 </Typography>
-              <Button variant='contained' onClick={() => removeItem(i)}>
+
+                <Typography variant={'h3'} color={'#333'} textAlign={'center'}>
+                      Quantity: {count}
+                </Typography>
+
+              <Button variant='contained' onClick={() => removeItem(name)}>
                 Remove
               </Button>
